Refatora skills de Mago para Multiclasse FIRE_BALL WATER_BUBBLE WIND_GUST EARTH_BREAK
